pyproject.toml
[project]
name = "drf-apischema"
version = "0.3.0"
description = "API schema generator and validator for Django REST framework"
readme = "README.md"
license = { file = "LICENSE" }
authors = []
requires-python = ">= 3.12"
keywords = [
    "drf",
    "django",
    "django-rest-framework",
    "schema",
    "swagger",
    "openapi",
    "documentation",
    "drf-spectacular",
    "drf-apischema",
]
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "django",
    "django-restframework",
    "rich",
    "sqlparser",
    "drf-spectacular",
]

[project.urls]
Repository = "https://github.com/hmeqo/drf-apischema.git"
Issues = "https://github.com/hmeqo/drf-apischema/issues"

[dependency-groups]
types = ["django-types", "djangorestframework-types"]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
